api/pyproject.toml
[tool.poetry]
name = "promptalchemy"
version = "0.1.0"
description = "PromptAlchemy is an open-source library designed to elevate your prompt engineering game. By applying advanced techniques like Chain-of-Thought, Few-Shot Learning, and more, we transform basic user prompts into highly effective, complex prompts."
authors = ["Minh-Duc Bui <minhduc2k31905@gmail.com>"]
readme = "README.md"
packages = [
    {include = "api/src"}
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10.14"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
pathlib = "^1.0.1"
pydantic = "^2.8.2"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"
langfuse = "^2.20.0"
fastapi = "0.111.0"
uvicorn = "0.22"
slowapi = "^0.1.9"
pydantic-settings = "^2.4.0"
isort = "^5.13.2"
redis = "^5.0.8"
litellm = {extras = ["proxy"], version = "^1.43.7"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
coverage = "~7.2.4"
pytest = "~8.1.1"
pytest-benchmark = "~4.0.0"
pytest-env = "~1.1.3"
pytest-mock = "~3.14.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = "~0.5.1"
dotenv-linter = "~0.5.0"

[tool.ruff]
exclude = [
]
line-length = 120
extend-include = ["*.ipynb"]

[tool.ruff.lint]
preview = true
select = [
    "B", # flake8-bugbear rules
    "C4", # flake8-comprehensions
    "F", # pyflakes rules
    "I", # isort rules
    "UP", # pyupgrade rules
    "B035", # static-key-dict-comprehension
    "E101", # mixed-spaces-and-tabs
    "E111", # indentation-with-invalid-multiple
    "E112", # no-indented-block
    "E113", # unexpected-indentation
    "E115", # no-indented-block-comment
    "E116", # unexpected-indentation-comment
    "E117", # over-indented
    "RUF019", # unnecessary-key-check
    "RUF100", # unused-noqa
    "RUF101", # redirected-noqa
    "S506", # unsafe-yaml-load
    "SIM116", # if-else-block-instead-of-dict-lookup
    "SIM401", # if-else-block-instead-of-dict-get
    "SIM910", # dict-get-with-none-default
    "W191", # tab-indentation
    "W605", # invalid-escape-sequence
    "F601", # multi-value-repeated-key-literal
    "F602", # multi-value-repeated-key-variable
]
ignore = [
    "F403", # undefined-local-with-import-star
    "F405", # undefined-local-with-import-star-usage
    "F821", # undefined-name
    "F841", # unused-variable
    "UP007", # non-pep604-annotation
    "UP032", # f-string
    "B005", # strip-with-multi-characters
    "B006", # mutable-argument-default
    "B007", # unused-loop-control-variable
    "B026", # star-arg-unpacking-after-keyword-arg
#    "B901", # return-in-generator
    "B904", # raise-without-from-inside-except
    "B905", # zip-without-explicit-strict
]

[tool.ruff.lint.per-file-ignores]
"app.py" = [
    "F401", # unused-import
    "F811", # redefined-while-unused
]
"__init__.py" = [
    "F401", # unused-import
    "F811", # redefined-while-unused
]
"tests/*" = [
    "F401", # unused-import
    "F811", # redefined-while-unused
]

[tool.ruff.format]
quote-style = "single"








[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
Sphinx = "~=7.0"
